backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
cachetools==6.2.1
    # via -r requirements.in
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via click
ecdsa==0.19.1
    # via python-jose
exceptiongroup==1.3.0
    # via anyio
fastapi==0.121.2
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.12.4
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via -r requirements.in
python-jose==3.5.0
    # via -r requirements.in
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
starlette==0.49.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.38.0
    # via -r requirements.in
